requirements.txt
flask
numpy
scikit-learn
pillow
plotly
dash
pandas